backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
minio==7.2.0
moviepy==1.0.3
Pillow==10.1.0
librosa==0.10.1
aiofiles==23.2.1
python-dotenv==1.0.0
httpx==0.25.2
websockets==12.0
